完善SpEL表达式&&修改缓存删除的逻辑，存在才删除

+++ double-cache-web/src/main/java/com/cqupt/study/aspect/DoubleCacheDeleteAspect.java
@@ -1,16 +1,28 @@
 package com.cqupt.study.aspect;
 
 import com.cqupt.study.annotation.DoubleCacheDelete;
+import com.cqupt.study.exception.ErrorException;
 import com.cqupt.study.guava.cache.StudentGuavaCache;
 import com.cqupt.study.pojo.Student;
 import com.cqupt.study.redis.RedisService;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
+import org.aspectj.lang.reflect.MethodSignature;
+import org.springframework.core.LocalVariableTableParameterNameDiscoverer;
+import org.springframework.expression.EvaluationContext;
+import org.springframework.expression.Expression;
+import org.springframework.expression.ExpressionParser;
+import org.springframework.expression.spel.standard.SpelExpressionParser;
+import org.springframework.expression.spel.support.StandardEvaluationContext;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
+import java.lang.reflect.Method;
+import java.util.Arrays;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * @Description:  查找双重缓存注解的切面
@@ -22,6 +34,16 @@ import javax.annotation.Resource;
 @Aspect
 @Component
 public class DoubleCacheDeleteAspect {
+    /**
+     * 表达式解析器
+     * */
+    ExpressionParser parser = new SpelExpressionParser();
+
+    /**
+     * 获取方法参数
+     * */
+    LocalVariableTableParameterNameDiscoverer discoverer = new LocalVariableTableParameterNameDiscoverer();
+
     @Resource
     private StudentGuavaCache studentGuavaCache;
 
@@ -39,12 +61,34 @@ public class DoubleCacheDeleteAspect {
     @Transactional(rollbackFor = Exception.class)
     public Object dealProcess(ProceedingJoinPoint pjd, DoubleCacheDelete doubleCacheDelete) {
         Object result = null;
-        String key = doubleCacheDelete.key();
+        Method method = ((MethodSignature) pjd.getSignature()).getMethod();
+
+        //获得参数名
+        String[] params = discoverer.getParameterNames(method);
+
+        //获得参数值
+        Object[] object = pjd.getArgs();
+
+        EvaluationContext context = new StandardEvaluationContext();
+        for (int i = 0; i < params.length; i++) {
+            context.setVariable(params[i], object[i]);
+        }
+
+
+        //解析SpEL表达式
+        if (doubleCacheDelete.key() == null) {
+            throw new ErrorException("@DoubleCacheDelete注解中key值定义不为null");
+        }
+
+        String key = parser.parseExpression(doubleCacheDelete.key()).getValue(context, String.class);
+        System.out.println("key: " + key);
         if (key != null) {
             //1.清除guava cache缓存
             studentGuavaCache.invalidateOne(Long.valueOf(key));
             //2.清除redis缓存
             redisService.delete(Long.valueOf(key));
+        } else {
+            throw new ErrorException("@DoubleCacheDelete注解中key值定义不存在，请检查是否和方法参数相同");
         }
 
         //执行目标方法
